adbms_ecommerce/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pymongo==4.9.1
pydantic==2.9.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
lxml==5.3.0
streamlit==1.38.0
requests==2.32.3
plotly==5.24.1
pandas==2.2.3
